New plugin: DynamicTapDance
This implements an extension to the `TapDance` plugin, allowing us to store
`tapDanceActionKeys()`-esque lists in Storage. The core idea here is very
similar to that of `DynamicMacros`: we dump/restore the full list via Focus,
build a cached index on setup and any updates, and play back the selected key
when need be.

Unlike `DynamicMacros`,this plugin is built on top of `TapDance` and cannot
function without it.

Fixes #730.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>