Fixed missing instanciation of driver::leds::Base<...>::LEDs_
This static member did not have an instance under
certain circumstances. A typical workaround for such
a situation is to make it a local static of an accessor
function. By this means the one definition rule is
not violated and the object always instanciated. It is still optimized
away by the compiler in case device::Base<...>::LEDs() is not called.

Signed-off-by: Florian Fleissner <florian.fleissner@inpartik.de>