device/Raise: Fix the column loop in actOnMatrixScan()
`RaiseKeyScanner::actOnMatrixScan()` erroneously looped through all the columns
in the matrix, while we only wanted to loop through the left half, since we read
the right half at the same time, not separately. This resulted in half the keys
producing two events, due to overflow. We also calculated the key number
incorrectly.

Both of these issues are fixed with this patch.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>