Add a way to override the HID short name
With this change, each device can specify a short name in their device
properties, which will be used to override the HID shortname. Due to link order,
we need to do the override in the user sketch, so we hook into the `KEYMAPS`
macro, to call `_INIT_HID_GETSHORTNAME`, which will set the override up for us.

The short name defaults to "kaleidoscope".

Together with keyboardio/KeyboardioHID#61, fixes keyboardio/KeyboardioHID#54.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>