examples/AppSwitcher: Do not use dynamic  vars in MACRO()
The `MACRO(...)` macro stores the macro in progmem, and as such, needs to use
constants only. Rework the logic in `macroAppSwitch()` so that we keep that
requirement.

For some reason, the previous version of the code worked with gcc 5.4, but not
with gcc 7+, and the latter is right to complain.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>