Update for latest firmware
Whether to drop the pixels in a given cycle and how much to decay the
intensities are now calculated based on time elapsed rather than by
incrementing a counter.

Because of this, configuration has changed.

Instead of DROP_TICKS, there is now DROP_MS.

Additionally, there is now DECAY_MS to control the time it takes for a
pixel to decay from full intensity to nothing.