examples/Atreus2: Let TG(UPPER) go back to QWERTY from UPPER
To activate the `Upper` layer, one has to press `Fun+Esc`, and layers get locked
to `Upper`. To go back to base, one would need to press `Fun`, while the `Upper`
key itself is disabled.

For the sake of laziness, this little change turns that disabled key on the
`Upper` layer into the same macro key as `Fun`, so hitting the same position
again will get us back to the QWERTY layer.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>