Implement a LayerSelect plugin
The new plugin gives us a new key, that works similarly to modifiers, but for
layers: hold it, and it has the chance to turn other keys into layer moving
keys. By default, the numbers between 0 and 9 will be treated as triggers, but
the method can be replaced with custom decision making.

Fixes #804.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>