Layer.move: Forcibly update all our caches
`Layer.move()` relied on `activate()` to update the caches, but `activate()`
worked under the assumption that we also `deactivate()` layers - which we did
not in `move()`, as we directly modified the state. Since we directly modify the
state, we can't rely on `activate()` either, and have to update the caches
ourselves.

Unlike `activate()`, we do this unconditionally, because there isn't a case
where we do not want to update them.

This makes `MoveToLayer()` play well with `Colormap`, and any other plugin that
rely on cached layer information, such as the topmost active layer.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>