Guard some places against interrupts
When checking or updating the key state, we need to do so with interrupts
disabled for the _whole_ duration of our operation, otherwise we may end up in
an inconsistent state. This means that we disable interrupts for the whole of
`actOnMatrixScan` (for hardware that uses interrupts for key scanning, anyway),
and whenever we use `pressedKeyswitchCount`, `previousPressedKeyswitchCount`,
`isKeyswitchPressed`, or `wasKeyswitchPressed`. These latter ones can't be
guarded within the functions, because we need to guard not just a single call to
them, but all of them when done in a loop.

Fixes #816, and also fixes #812 in a different, better way.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>