dygma/Raise: Adjust the red component of LED colors
It seems the red component in the LEDs is a tiny bit stronger than the others,
so lets adjust the component's value a little on the firmware side to make
colors come out right.

This is not the best solution, it should eventually be configurable, but until
then, this is the best workaround we could come up with.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>