Restore ATmega32U4Keyboard compatibility
When `FlashHelper` was implemented, we updated the `ATmega32U4Keyboard` (&
related) macros to require a prog key definition. This breaks compatibility with
any keyboard using the macros, and requires us to update them in lockstep. While
we can do that within Kaleidoscope itself, it sets a bad precedent.

As such, lets restore those macros to their original shape, and introduce a set
of new ones for `Prog` key support. This allows boards that do not have a prog
key to remain unmodified.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>