Future-proof our device & driver helper macros
We currently have `ATMEGA32U4_KEYBOARD` as a helper macro to make it easier to
create derived keyboards, but it takes a fixed, set amount of arguments, so
anytime we want to add a new device property, we'd have to create a new macro,
which doesn't scale.

So instead of having one macro with many arguments, lets introduce a different
kind of helper macro: one that abstracts away the boring parts, but still lets
you assemble the device descriptions piece by piece. This is more verbose,
indeed, but it is also clearer, and easier to extend. We can just add new
properties for devices that need it, and leave the rest alone without breaking
anything.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>