Update the udev rules to tag devices as uaccess and seat
In order to be able to access the devices as the at-seat user, without having to
fiddle with distro-specific permissions and groups, we need to tag it both
`uaccess` and `seat`, and have the rule sorted before the one that applies
permissions based on these tags. As such, the file had to be renamed as well.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>