kbdfans/KBD4x: Stop using opaque macros
Instead of using opaque macros which aren't even extensible, just expand them,
and use the raw data structures for the hardware definition. While this is more
verbose than the macros, it is more future proof, and clearer for the reader
too, because they don't need to understand the magic macros.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>