kaleidoscope-builder: Allow easily selecting a compiler
While we do not currently support any other compiler than the gcc suite, we
eventually want to. Meanwhile, we also want to be able to easily switch to gcc
when that's not the default on the system.

For this reason, we introduce `KALEIDOSCOPE_COMPILER`, which can be set to `gcc`
or `clang`, and `kaleidoscope-builder` will then set up the appropriate
basenames.

This, combined with the previous change, allows us to easily select gcc for
virtual builds, even when clang is the default on the system.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>