TapDance: Do not mask interrupting keys anymore
We introduced the masking to avoid sending extra keys when the mapped key
changes prior to release - but since the introduction of the caching mechanism,
we no longer need to do this.

However, for the caching to work the way we want it to, we need to map the key
to `NoKey` once, upon interrupting.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>

squash! TapDance: Do not mask interrupting keys anymore

Signed-off-by: Gergely Nagy <algernon@keyboard.io>