Deprecate the key masking API
Key masking was a bandaid, and we have better ways to achieve the same thing
now. All current users have been switched over to different methods now, so lets
deprecate the masking.

We only put the `DEPRECATED` label on the `maskKey` method, because the rest are
used internally too, and we do not want to emit warnings for those.

Fixes #884.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>