Simplify Escape-OneShot
I removed the unnecessary `did_escape_` boolean, and instead changed
Escape-OneShot so that it is only triggered if the `escape` key has
toggled on.

Additionally, I changed the conditions to agree better with the
docs. Previously, if there was a sticky OneShot key, the `escape` key
would cancel OneShots even if a OneShot key was being held, contrary
to the docs. I suspect this was a logic error in the rather tricky
conditional that was being used.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>