Rewrite OneShot plugin
This is a major rewrite of the OneShot plugin to make it more
efficient and reliable. The event handler has been reorganized, the
other hooks simplified, and a few helper functions added.

One change of note is the addition of the `replaceOneShot()` helper
function, which temporarily replaces a held OneShot key with its
corresponding modifier or layer shift key in the keymap. This is
called when a OneShot key has been cancelled, but is still being
held.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>