Remove all calls to `replaceOneShot()`
I tried rewriting replaceOneShot() to fix a problem with OSL
keys. A OneShot layer that has another key mapped to its position on
the target layer gets clobbered by that key on a double-tap, and
so can't become sticky, or get cancelled early by a second tap. The
trouble is that this isn't fixable from within the plugin itself.

Also added lots of explanatory comments.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>