Include pressed OneShot keys as active in `isActive(key)`
This introduces a discrepancy between the OneShot `isActive()` global
function and the keycode-specific one. I did not include pressed keys
in the global version because that one seems better suited to letting
other plugins know if there's a chance a OneShot key is active despite
not being held. The key-specific one seems more suitable to including
the `pressed` but not `active` state, because a held OneShot key very
quickly transitions from one to the other after `hold_time_out` passes.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>