Add a small tool, to run things within Docker
To make it easier to reproduce things, and to help build in a clean environment,
this adds a thin Dockerfile that has Arduino and arduino-cli pre-installed, and
- along with the `bin/run-docker` script - is set up so that one can easily run
arbitrary commands in the context of the current bundle and Kaleidoscope.

The first run will take a while, because docker will build the image. Subsequent
runs will use the cache.

To use: `bin/run-docker make`, for example. Any argument passed to the
`bin/run-docker` will be eval-ed within the container, and will run there.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>