device/dygma/raise: Rename raise::Hand to raise::RaiseSide
We're going to merge KeyboardioScanner into Kaleidoscope as a driver, but we
can't name that model01::Hand, because then we have a filename conflict due to
our use of static archiving during linking. To avoid that, both will use a
similar, but unique naming pattern, and raise::Hand becomes raise::RaiseSide
instead.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>