Use eager reporting when debouncing
If a key has been in a stable state and then changes, report that change
immediately and then wait for a cooldown period before reporting any
additional changes.  This reduces latency for registering initial
keypresses.

It would be unusual for this to introduce chatter, since mechanical
switches only tend to be noisy immediately after a key press or release.

This change does not introduce any additional persistent state, although
it does add an additional parameter to the debouncer to distinguish
between the last raw state and the last debounced state.

It also fixes a bug whereby all rows in the matrix would be rewritten
for every row that changed due to a mis-nesting of loops.  However,
it now writes to the matrix state unconditionally.

Signed-off-by: Dean Scarff <dos@scarff.id.au>