Make ActiveModColor work if registered ahead of OneShot
If `ActiveModColorEffect` was registered ahead of `OneShot` in
`KALEIDOSCOPE_INIT_PLUGINS()`, `OSM()` and `OSL()` keys would light up in the
OneShot "sticky" state, not in the "held" or "one-shot" states. This happened
because OneShot changes the `event.key` value to the corresponding base
key (modifier or layer shift), but if ActiveModColor had already processed that
key event, it wouldn't recognize the key as a modifier/layer shift key, and
would therefore ignore it.

This change makes ActiveModColor also recognize OneShot keys as modifier/layer
shift keys.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>