Add `afterReportingState()` event handler function
This event handler is useful for plugins that need to react to events, but
should wait until after those events are fully processed before doing so.  This
is useful for OneShot, which needs to keep keys active until after events that
trigger their release.  The `afterEachCycle()` hook is unfortunately
insufficient for this purpose, because the same event could trigger multiple
plugins (e.g. TapDance & OneShot) to resolve events, and the OneShot should
apply only to the first ensuing report.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>