Add `afterReportingState()` handler to OneShot
If multiple events are processed in a single cycle, we want a OneShot key whose
release is triggered by the first one to only affect that key, and not
subsequent ones.  For example, if we tap `OSM(LeftShift)`, then `TD(0)`, then
`Key_X`, the OneShot shift should only apply to the output of the TapDance key,
not the `x`.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>