Add a "no-delay" mode to SpaceCadet
SpaceCadet now has three "modes": on, off, and on with no delay. In "no-delay"
mode, when a SpaceCadet key is pressed, its primary (modifier) key value is sent
immediately to the host, and if it is released before timing out, that value is
then replaced by the configured "tap" value of the key.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>