Change default value of `KeyEvent` from `Key_NoKey` to `Key_Undefined`
This lets us remove some awkward code from `handleKeyswitchEvent()`, because as
long as the default value (which triggers a keymap lookup) was the same as
`Key_Masked`, it wasn't sufficient for an `onKeyswitchEvent()` handler to change
`event.key` to `Key_Masked`, because that would be interpreted by
`handleKeyEvent()` as a signal to do a keymap lookup.

This also makes it more consistent with other parts of the code.  The values
`Key_Undefined`, `Key_Inactive`, and `Key_Transparent` are all the same, and
with this change they are each interpreted the same way by code that encounters
them.  In a keymap lookup, if an active layer has a `Key_Transparent` value, we
continue searching for a different value on another layer.  In the live keys
array, if we find a `Key_Inactive` entry, we look for a value from the keymap.
And with this change, when processing a key event, if it has a `Key_Undefined`
value, we look for a value from `active_keys` (and then from the keymap layers,
if nothing is found there).

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>