EscapeOneShot: Make it possible to disable the plugin at run-time
We want to make it possible to have the plugin in firmwares shipped by
Chrysalis, but still have the functionality optional. To achieve this, we need
to be able to toggle it on and off at will.

We move both the existing `cancel_oneshot_key_` property, and the new toggle
into a struct, which we will later make use of in the upcoming configuration
plugin.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>