Model100: Use the real Serial, instead of the default NoSerial
By default, `kaleidoscope::device::Base<>` uses a dummy serial implementation,
and devices must override that to use the proper one. For the Model01, on which
the Model100 plugin was based on, this is done by the `ATMega32U4Keyboard`
class. For the Model100, we need to do it ourselves, because we're deriving from
Base directly.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>