Remove deprecated `handleKeyswitchEvent()` function
This removes the `key_events.*` files that once contained the main
`handleKeyswitchEvent()` function, and all references to it.  Because
`key_events.h` was included in the main `Kaleidoscope.h` header file,
`key_defs.h` and `keyswitch_state.h` were added to that header so that other
code that relies on those things being included via `Kaleidoscope.h` will
continue to work.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>