Simplify our docker execution (#1124)
* Switch docker to no longer build a bundle separate from the kaleidoscope
dir.
Stop copying over a bunch of tarballs we don't need.

* fix shellcheck

* Stop copying the core-bundled copies of Kaleidoscope over to docker

* Symlink our Kaleidoscope into the versions packaged into the arduino
cores when running in docker. We need to do this to satisfy the
sketch_header.h/sketch_footer.h system we use to be able to amend
arduino sketches on avr and virtual. (This feature is not used by the
core and may be removed in the future)

* Reimplement docker-clean to do much less work and get the same result

* Add a stub at docs for the docker test runner