Correct Mouse HID `click()` method
This method was only sending one report, but two are necessary for a "click".
There was an assumption, based on the old Kaleidoscope, that a report would be
sent every cycle, so the release report would be "free".  This is not true of
Kaleidoscope any longer, and any other client might also not behave that way.

Also, if the mouse button was already active ("held"), the "press" report would
be a no-op, so instead of a click, we'd just get a release.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>