Rewrite MouseKeys plugin
This is a major rewrite of the MouseKeys plugin, primarily focused on mouse
cursor movement keys.  There is no change to the keys themselves, and the
behaviour is fairly similar, but there are now better configuration parameters
and defaults.  The new parameters are a minimum speed, a maximum speed, and the
length of time it takes to reach that maximum speed, without needing to worry
about the report update interval.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>